Added an algorithm to break up paths by "contiguous" points
I set the default "break" timer to 5 minutes, meaning that any GPS data points longer apart than 5 minutes are not counted as being part of the same path.  A new path is started in this case.  If a point has more than 5 minutes before and after it, it is discarded and no path is made.